pyproject.toml
[tool.poetry]
name = "boths"
version = "0.1.0"
description = "Telegram to help manage Hackerspace"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.1"
requests = "^2.25"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
jedi = "^0.17.2"
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
